output/prototypes/uni1_image_ad/requirements.txt
requests>=2.28.0
pyyaml>=6.0
Pillow>=9.0
